feat: Introduce unified interface for CommonMark and HTML output
- Added a new example demonstrating the use of `to_commonmark` and `to_html` interfaces.
- Implemented block-level element writing functionality in `block.rs`.
- Added HTML fallback handling for tables with block elements in `html_fallback.rs`.
- Developed inline element writing functionality in `inline.rs`.
- Organized CommonMark writer into modular components in `mod.rs`.
- Implemented table writing functionality in `table.rs`.
- Created utility functions and character escaping in `utils.rs`.
- Enhanced the main CommonMark writer functionality in `writer.rs`.
