+++ src/ast/mod.rs
@@ -3,12 +3,10 @@
 //! This module defines various node types for representing CommonMark documents,
 //! including headings, paragraphs, lists, code blocks, etc.
 
-mod custom;
 mod html;
 mod node;
 pub mod tables;
 
-pub use self::custom::{GenericCustomNode, NodeKind, TextCustomNode};
 pub use self::html::{HtmlAttribute, HtmlElement};
 pub use self::node::{CodeBlockType, HeadingType, ListItem, Node};
 pub use crate::traits::CustomNode;
