+++ src/writer/mod.rs
@@ -2,7 +2,7 @@
 //!
 //! This module provides functionality to convert AST nodes to various formats.
 
-mod cmark;
+pub mod cmark;
 pub mod processors;
 
 pub use self::cmark::CommonMarkWriter;
